refactor(app): move healthcheck handler and test to internal/app/handler.go
