+++ internal/app/handlers.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"fmt"
@@ -7,8 +7,8 @@ import (
 
 // Declare a handler which writes a plain-text response with information about the
 // application status, operating environment and version.
-func (app *application) healthcheckHandler(w http.ResponseWriter, r *http.Request) {
+func (app *Application) healthcheckHandler(w http.ResponseWriter, r *http.Request) {
 	fmt.Fprintln(w, "status: available")
-	fmt.Fprintf(w, "environment: %s\n", app.config.Env)
-	fmt.Fprintf(w, "version: %s\n", version)
+	fmt.Fprintf(w, "environment: %s\n", app.Config.Env)
+	fmt.Fprintf(w, "version: %s\n", app.Version)
 }
